requirements.txt
requests==2.32.3
newspaper3k==0.2.8
google-generativeai==0.8.3
pillow==9.0.0
tweepy==4.14.0
lxml[html_clean]
